claude-test/requirements.txt
Flask>=2.0.0
Flask-CORS>=4.0.0
google-generativeai
PyPDF2>=3.0.0
python-docx>=0.8.0
Werkzeug>=2.0.0